chore: refactor and update deps
