+++ .gitignore
@@ -3,3 +3,5 @@ build/
 dist/
 
 .ruff_cache/
+
+.vscode/
