+++ pyproject.toml
@@ -13,10 +13,11 @@ title-belt-nhl = 'title_belt_nhl.main:cli'
 [tool.poetry.dependencies]
 python = "^3.10"
 click = "^8.1.7"
+ruff = "0.6"
 
 
 [tool.poetry.group.dev.dependencies]
-ruff = "^0.1.5"
+ruff = ">=0.6.0"
 
 [build-system]
 requires = ["poetry-core"]
